+++ tests/testthat/test-environment-issues.R
@@ -34,15 +34,16 @@ test_that("send inquiry to estimator works", {
 })
 
 
-# To do: address warning
-test_that("no pars summary", {
+test_that("no parameters output", {
 
     design <-
     declare_model(N = 5, noise = rnorm(N)) +
     declare_inquiry(mean_noise = mean(noise))
 
-    DeclareDesign:::find_all_objects(design)
+    txt <- capture.output(print(DeclareDesign:::find_all_objects(design)))
+    expect_true(any(grepl("No parameters", txt, fixed = TRUE)))
     
-    expect_warning(redesign(design, N = 6)) 
+    expect_warning(redesign(design, N = 6), 
+                   "You requested a change to N but N is not found in the design") 
 
 })