+++ tests/testthat/test-redesign.R
@@ -1,5 +1,89 @@
 context("redesign")
 
+test_that("modify many steps", {
+  
+  a = 1
+  b = 2
+  c = 3
+  f = 4
+
+  design <- 
+    declare_model(N = a, x1 = b) + 
+    declare_measurement(x2 = b, x3 = c) +
+    declare_inquiry(
+      q1 = mean(x1),
+      q2 = mean(x2),
+      q3 = mean(x3),
+      q4 = f
+    )
+
+  design_2 <- redesign(design, a = 2, b = 1, c = 4, f = 5) 
+  
+  expect_true(all((design |> draw_estimands())$estimand == c(2,2,3,4)))
+
+  expect_true(all((design_2 |> draw_estimands())$estimand == c(1,1,4,5)))
+  
+})
+
+
+test_that("changing handlers", {
+  
+  x <- 2
+  
+  my_handler <- function(N) data.frame(X = 1:2)
+  
+  design <-
+    declare_model(
+      handler = my_handler) +
+    declare_model(X = x)
+  
+  DeclareDesign:::find_all_objects(design)
+  
+  my_handler_2 <- function(N) data.frame(X = 1:3)
+  
+  design_2 <- DeclareDesign:::modify_edit(design, my_handler = my_handler_2)
+  
+  expect_true(
+    draw_data(DeclareDesign:::redesign(design, my_handler = my_handler_2)) |> nrow() == 3
+  )
+  
+})
+
+
+test_that("modify functions", {
+  
+  b = -1
+  a = .1
+  f = function(x, a) a*x + b
+  
+  design <- 
+    declare_model(N = 1, x = 0) + 
+    declare_measurement(Y = f(x, a))
+  
+  design_2 <- redesign(design, f = function(x, a) 7)
+  
+  expect_true((design_2 |> draw_data())$Y == 7)
+  
+})
+
+
+test_that("modify data", {
+  
+  df <- data.frame(X = 1)
+  
+  design <- 
+    declare_model(data = df) + 
+    declare_measurement(Y = 2 * X)
+
+  design_2 <- DeclareDesign:::modify_edit(design, df = data.frame(X = 1:2, Z = 3))
+  expect_true(design_2 |> draw_data() |> nrow() == 2)
+
+  design_2 <- redesign(design, df = list(data.frame(X = 1:2, Z = 3)))
+  expect_true(design_2 |> draw_data() |> nrow() == 2)
+  
+})
+
+
 test_that("N changing", {
   N <- 100
   design <- declare_model(N = N) + NULL
@@ -25,18 +109,47 @@ test_that("N changing", {
 })
 
 
-test_that("message if meaningless change", {
-  N <- 100
-  design <- declare_model(N = N) + NULL
-  expect_message(redesign(design, k = 2, N = 200), "You requested a change to k but k is not found in the design")
+test_that("modify inside a handler", {
+  
+  m <- 2
+  b <- 100
+  N <- pi  # distraction
+  f <- function(...) fabricate(...)
+  
+  hdl <- function(...) f(..., extra = rnorm(N, b, 0))
+  
+  design <- 
+    declare_model(N = m, U = rnorm(N)) + 
+    declare_measurement(handler = hdl)
+  
+  rm(N, b, f, hdl)
+  
+  obs <- DeclareDesign:::find_all_objects(design)
+  # Rm N
+  expect_true(all(obs$name == c("m", "b", "f", "hdl")))
+  expect_true(mean(draw_data(design)$extra) == 100)
+  
+  
+  design <- design|> redesign(b = -100)
+
+  design_2 <- DeclareDesign:::modify_edit(design, b = -100)
+  obs <- DeclareDesign:::find_all_objects(design_2)
+  
+  expect_true(mean(draw_data(design_2)$extra) == -100)
+  
 })
 
+
+test_that("warning if meaningless change", {
+  N <- 100
+  design <- declare_model(N = N) + NULL
+  expect_warning(redesign(design, k = 2, N = 200), "You requested a change to k but k is not found in the design")
+})
 
 
 test_that("embedded parameter, multiple steps", {
   n <- 5000
   b <- .2
-  
   d <- 
     declare_model(N = n) + 
     declare_model(Y = rnorm(N, b)) + 
@@ -45,20 +158,14 @@ test_that("embedded parameter, multiple steps", {
   
   expect_true(run_design(d)$estimand == .2)
   rm(b)
-    expect_true(run_design(d)$estimand == .2)
-  
-  DeclareDesign:::find_all_objects(d)
-  d
+  expect_true(run_design(d)$estimand == .2)
   
   d2 <- redesign(d, b = .4)
-  DeclareDesign:::find_all_objects(d2)
-  d2
-    
+
   x <- run_design(d2)
   expect_true(x$estimand == .4)
   expect_true(abs(x$estimand - x$estimate) < .4)
-  attr(d2[[1]], "dots")
-  
+
 
   d3 <- redesign(d, b = 10)
   x <- run_design(d3)
@@ -69,135 +176,53 @@ test_that("embedded parameter, multiple steps", {
 
 
 
-
-test_that("embedded parameter, multiple steps", {
-  n <- 2000
-  b <- .2
+test_that("data_step OK", {
   
-  d <- 
-    declare_model(N = n,  Y = 1) + 
-    declare_estimator(Y ~ 1) +
-    declare_inquiry(Q = b)
-  
-  expect_true(run_design(d)$estimand == .2)
-  
-  rm(b)
-  
-  expect_true(run_design(d)$estimand == .2)
-  
-  DeclareDesign:::find_all_objects(d)
-  d
-  
-  d2 <- redesign(d, b = .4)
-  DeclareDesign:::find_all_objects(d2)
-  d2
-
-  # N is found after redesgin
-  b <- 2
-  simple_1 <- declare_model(N = 3,  Y = N + b) + NULL 
-  simple_1[[1]]()
-  
-  simple_2 <- DeclareDesign:::modify_edit(simple_1, b = 3)
-  simple_2[[1]]()
-
-
-  
-  
-
-  # N is found after redesgin
-  n <- 3
-  b <- 2
-  simple_1 <- declare_model(N = n,  Y = rnorm(N, b)) + NULL 
-  simple_1[[1]]()
-  
-  simple_2 <- DeclareDesign:::modify_edit(simple_1, b = 3)
-  simple_2[[1]]()
-
-  simple_3 <- DeclareDesign:::redesign(simple_1, b = 3)
-  simple_3[[1]]()
-  
-
-
-  #  OK
-  d <- 
-    declare_model(N = n, Y = rnorm(N, b)) + 
-    NULL
-  
-  d2 <- redesign(d, b = .4)
-  d[[1]]()
-  d2[[1]]()
-
-
-  # Not OK
-  d <- 
-    declare_model(N = n, Y = rnorm(N, b)) + 
-    declare_inquiry(Q = b)
-  
-  d2 <- redesign(d, b = .4)
-  d[[1]]()
-
-
-  expect_error(d2[[1]](), NA)
-  
-
-  
-})
-
-
-test_that("data_step throwing error", {
-  n <- 2000
+  n <- 1
   b <- .2
 
   d <- 
-    declare_model(N = n) + 
-    declare_inquiry(Q = b)
+    declare_model(N = n, Y = b) + 
+    declare_inquiry(Q = b) 
+    
   
-  redesign(d, b = .4)
-  
-  d <- 
-    declare_model(N = n, Y = rnorm(N, b)) + 
-    declare_inquiry(Q = b)
-  
-  d2 <- redesign(d, b = .4)
+  d <- redesign(d, b = .3)
 
-  attr(d[[1]], "dots")
-  attr(d2[[1]], "dots")
-  
-  expect_error(d2[[1]](), NA)
+  expect_true(d[[1]]()$Y == .3)
 
 })
 
 
-test_that("to do: changes inside functions", {
+test_that("future: changes inside functions", {
 
   b = -1
   a = .1
   f = function(x, a) a*x + b
 
   design <- 
-    declare_model(N = 2, x = runif(N)) + 
+    declare_model(N = 1, x = 0) + 
     declare_measurement(Y = f(x, a))
   
   DeclareDesign:::find_all_objects(design)
   
   rm(b, a, f)
   
-  design |> draw_data()
+  expect_true((design |> draw_data())$Y == -1)
 })
 
 
-test_that("to do: changing handlers", {
-  
-  my_handler <- fabricate
-  
-  design <-
-    declare_model(
-      handler = my_handler,
-      N = 5,
-      Y = rnorm(N, sd = 0.25)
-    ) +
-    declare_inquiry(Y_bar = mean(Y)) 
 
-  DeclareDesign:::find_all_objects(design)
+
+
+test_that("future: modify d; avoid partial matching problems", {
   
+  d = 1
+
+  design <- 
+    declare_model(N = 1, D = d) + NULL
+
+  expect_error(DeclareDesign:::modify_edit(design = design, d=2), NA) 
+  expect_error(redesign(design = design, d=2)) 
+  
+
 })